@@ -21,17 +21,16 @@ use Zend\Form\Form;
 use Zend\InputFilter\InputFilterProviderInterface;
 use Zend\Validator\File\MimeType;
 use Zend\Validator\File\Size;
+use Zend\Form\Element\Submit;
+use Zend\Form\Element\File;
 
 /**
  * Class Import
  *
  * @package Organisation\Form
  */
-class Import extends Form implements InputFilterProviderInterface
+final class Import extends Form implements InputFilterProviderInterface
 {
-    /**
-     * Import constructor.
-     */
     public function __construct()
     {
         parent::__construct();
@@ -41,30 +40,30 @@ class Import extends Form implements InputFilterProviderInterface
 
         $this->add(
             [
-                'type'    => '\Zend\Form\Element\File',
+                'type'    => File::class,
                 'name'    => 'file',
                 'options' => [
-                    "label" => "txt-file",
+                    'label' => 'txt-file',
                 ],
             ]
         );
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Submit',
+                'type'       => Submit::class,
                 'name'       => 'upload',
                 'attributes' => [
-                    'class' => "btn btn-primary",
-                    'value' => _("txt-upload"),
+                    'class' => 'btn btn-primary',
+                    'value' => _('txt-upload'),
                 ],
             ]
         );
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Submit',
+                'type'       => Submit::class,
                 'name'       => 'import',
                 'attributes' => [
-                    'class' => "btn btn-primary",
-                    'value' => _("txt-import"),
+                    'class' => 'btn btn-primary',
+                    'value' => _('txt-import'),
                 ],
             ]
         );
